requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-unique-user-email
djangorestframework
djangorestframework-simplejwt
EasyProcess
entrypoint2
filelock
Markdown
MouseInfo
mss
numpy
opencv-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
setuptools
sqlparse
typing_extensions
tzdata
virtualenv